Small adjustments. Use -O3 for release builds.

+++ src/kernel_info.cpp
@@ -56,7 +56,7 @@ std::tuple<dim3d, dim3d> get_block_and_grid_size_reduction(const Policy& policy)
     const unsigned block_size = reduction.local_block_size(functor);
 
     dim3d block{1, block_size, 1};
-    dim3d grid{std::min(block[1], nb_work), 1, 1};
+    dim3d grid{std::min((int) block[1], nb_work), 1, 1};
 
     return std::make_tuple(block, grid);
 }
@@ -77,7 +77,7 @@ std::tuple<dim3d, dim3d> get_block_and_grid_size(const Policy& policy)
 
     const unsigned block_size = Kokkos::Experimental::Impl::hip_get_preferred_blocksize<ParallelFor, LaunchBounds>();
     const dim3d block{1, block_size, 1};
-    const dim3d grid{typename Policy::index_type((nb_work + block[1] - 1) / block[1]), 1, 1};
+    const dim3d grid{static_cast<unsigned int>(typename Policy::index_type((nb_work + block[1] - 1) / block[1])), 1, 1};
 
     return std::make_tuple(block, grid);
 }
@@ -99,7 +99,7 @@ std::tuple<dim3d, dim3d> get_block_and_grid_size_reduction(const Policy& policy)
     const unsigned block_size = reduction.local_block_size(functor);
 
     dim3d block{1, block_size, 1};
-    dim3d grid{std::min(block[1], typename Policy::index_type((nb_work + block[1] - 1) / block[1])), 1, 1};
+    dim3d grid{std::min((unsigned int)(block[1]), (unsigned int) (typename Policy::index_type((nb_work + block[1] - 1) / block[1]))), 1, 1};
 
     return std::make_tuple(block, grid);
 }
